Fixed eslint/prettier, TCCB, AuthScreens
