+++ package.json
@@ -1,22 +1,21 @@
 {
-"name": "chore-ify",
-"version": "1.0.0",
-"main": "node_modules/expo/AppEntry.js",
-"scripts": {
-	"start": "expo start",
-	"android": "expo start --android",
-	"ios": "expo start --ios",
-	"web": "expo start --web",
-	"dev": "next",
-	"build": "next build",
-	"startnext": "next start",
-	"style:all": "tsc && npm run style:lint && npm run style:prettier",
-	"style:lint": "eslint .",
-	"style:prettier": "prettier --check '**/*.{js,jsx,ts,tsx}'",
-	"make-pretty": "prettier --write '**/*.{js,jsx,ts,tsx}'",
-	"lint": "eslint ."
-},
-
+  "name": "chore-ify",
+  "version": "1.0.0",
+  "main": "node_modules/expo/AppEntry.js",
+  "scripts": {
+    "start": "expo start",
+    "android": "expo start --android",
+    "ios": "expo start --ios",
+    "web": "expo start --web",
+    "dev": "next",
+    "build": "next build",
+    "startnext": "next start",
+    "style:all": "tsc && npm run style:lint && npm run style:prettier",
+    "style:lint": "eslint .",
+    "style:prettier": "prettier --check '**/*.{js,jsx,ts,tsx}'",
+    "make-pretty": "prettier --write '**/*.{js,jsx,ts,tsx}'",
+    "lint": "eslint ."
+  },
   "dependencies": {
     "@react-navigation/material-bottom-tabs": "^6.2.17",
     "@react-navigation/material-top-tabs": "^6.6.4",
@@ -29,8 +28,8 @@
     "react-native-chart-kit": "^6.12.0",
     "react-native-pager-view": "^6.2.0",
     "react-native-paper": "^5.10.6",
-    "react-native-safe-area-context": "^4.7.2",
-    "react-native-svg": "^13.14.0",
+    "react-native-safe-area-context": "^4.6.3",
+    "react-native-svg": "13.9.0",
     "react-native-tab-view": "^3.5.2",
     "react-native-vector-icons": "^10.0.0",
     "react-navigation-stack": "^2.10.4",
@@ -42,16 +41,15 @@
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.14",
     "@types/react-native-vector-icons": "^6.4.15",
-    "typescript": "^5.1.3",
-	"@typescript-eslint/eslint-plugin": "^6.7.5",
-	"@typescript-eslint/parser": "^6.7.5",
-	"eslint": "^8.51.0",
-	"eslint-config-airbnb": "^19.0.4",
-	"eslint-config-airbnb-typescript": "^17.1.0",
-	"eslint-config-prettier": "^9.0.0",
-	"eslint-config-universe": "^12.0.0",
-	"eslint-plugin-prettier": "^5.0.1"
+    "@typescript-eslint/eslint-plugin": "^6.7.5",
+    "@typescript-eslint/parser": "^6.7.5",
+    "eslint": "^8.51.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-config-universe": "^12.0.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "typescript": "^5.1.3"
   },
-
-	"private": true
+  "private": true
 }
