+++ src/clickableComponents/ThemedClickableCardButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { TouchableOpacity, View } from 'react-native';
-import { Card, Paragraph } from 'react-native-paper';
+import { Card, Paragraph, Title } from 'react-native-paper';
 import Icon from 'react-native-vector-icons/FontAwesome';
 import { useAppTheme } from '../contexts/ThemeContext';
 
@@ -10,17 +10,25 @@ type CardButtonProps = {
   iconName: string;
   onPress: () => void;
   iconSize?: number;
+  hideTitle?: boolean; // Add this line
 };
 
 export default function ThemedClickableCardButton({
-  // title,
+  title,
   content,
   iconName,
   onPress,
   iconSize,
+  hideTitle,
 }: CardButtonProps) {
   const theme = useAppTheme(); // get the theme
   const actualIconSize = iconSize ?? 30; // Use 30 if `iconSize` is not provided
+  const renderTitle = () => {
+    if (!hideTitle) {
+      return <Title>{title}</Title>;
+    }
+    return null;
+  };
 
   return (
     <TouchableOpacity onPress={onPress}>
@@ -38,7 +46,7 @@ export default function ThemedClickableCardButton({
             style={{ marginRight: 15 }}
           />
           <View>
-            {/* <Title>{title}</Title> */}
+            {renderTitle()}
             <Paragraph style={{ color: theme.colors.primary }}>
               {content}
             </Paragraph>
