+++ tsconfig.json
@@ -1,7 +1,9 @@
 {
-	"extends": "expo/tsconfig.base",
-	"compilerOptions": {
-		"strict": true,
-		"jsx": "react"
-	}
+  "extends": "expo/tsconfig.base",
+  "compilerOptions": {
+    "strict": true,
+    "jsx": "react"
+  },
+  "include": ["src/**/*", "assets/**/*", "types/**/*", "App.tsx"],
+  "exclude": ["node_modules"]
 }
